+++ CHANGELOG.md
@@ -36,3 +36,4 @@
 - Do not access `PlannedStmt` after we call `standard_ProcessUtility()` ([PG-2486](https://perconadev.atlassian.net/browse/PG-2486))
 - Various improvements to the stability of our test suite
 - Make sure that for prepared statements utility statement exec info read at the executor start hook, where data is not yet modified by query itself
+- Do not acquire LWLock under spinlock